+++ test/helper.rb
@@ -1,6 +1,6 @@
 # https://coveralls.io Integration
 require 'coveralls'
-Coveralls.wear_merged!
+Coveralls.wear!
 
 # Gem
 require 'rend/acl'
@@ -12,8 +12,3 @@ require 'support/failing_assertion'
 
 # Autorun tests
 require "minitest/autorun"
-
-# Optionally try to use TURN gem if we're in a RUBY_VERSION >= 1.9.3 environment
-if Gem::Version.new(RUBY_VERSION) > Gem::Version.new('1.9.2')
-  begin; require 'turn/autorun'; rescue LoadError; end
-end